Add some tests, extract.
