+++ spec/fixtures/dspace/1672.json
@@ -1,0 +1,232 @@
+{
+  "id": 1672,
+  "name": "Travelling across dinner-tables.",
+  "handle": "88435/dsp012801pg38m",
+  "type": "item",
+  "link": "/rest/items/1672",
+  "expand": [],
+  "lastModified": "2012-10-07 02:01:57.234",
+  "parentCollection": {
+    "id": 222,
+    "name": "Monographic reports and papers (Publicly Accessible)",
+    "handle": "88435/dsp016q182k16g",
+    "type": "collection",
+    "link": "/rest/collections/222",
+    "expand": [
+      "parentCommunityList",
+      "parentCommunity",
+      "items",
+      "license",
+      "logo",
+      "all"
+    ],
+    "logo": null,
+    "parentCommunity": null,
+    "parentCommunityList": [],
+    "items": [],
+    "license": null,
+    "copyrightText": "",
+    "introductoryText": "",
+    "shortDescription": "",
+    "sidebarText": "",
+    "numberItems": 1982
+  },
+  "parentCollectionList": [
+    {
+      "id": 222,
+      "name": "Monographic reports and papers (Publicly Accessible)",
+      "handle": "88435/dsp016q182k16g",
+      "type": "collection",
+      "link": "/rest/collections/222",
+      "expand": [
+        "parentCommunityList",
+        "parentCommunity",
+        "items",
+        "license",
+        "logo",
+        "all"
+      ],
+      "logo": null,
+      "parentCommunity": null,
+      "parentCommunityList": [],
+      "items": [],
+      "license": null,
+      "copyrightText": "",
+      "introductoryText": "",
+      "shortDescription": "",
+      "sidebarText": "",
+      "numberItems": 1982
+    }
+  ],
+  "parentCommunityList": [
+    {
+      "id": 146,
+      "name": "Princeton University Library",
+      "handle": "88435/dsp01g732d9019",
+      "type": "community",
+      "link": "/rest/communities/146",
+      "expand": [
+        "parentCommunity",
+        "collections",
+        "subCommunities",
+        "logo",
+        "all"
+      ],
+      "logo": null,
+      "parentCommunity": null,
+      "copyrightText": "",
+      "introductoryText": "",
+      "shortDescription": "",
+      "sidebarText": "",
+      "countItems": 12954,
+      "subcommunities": [],
+      "collections": []
+    }
+  ],
+  "metadata": [
+    {
+      "key": "dc.contributor.author",
+      "value": "Grigorieva, Alexandra",
+      "language": null
+    },
+    {
+      "key": "dc.contributor.author",
+      "value": "Antonov, Alexandr",
+      "language": null
+    },
+    {
+      "key": "dc.date.accessioned",
+      "value": "2012-10-01T18:08:11Z",
+      "language": null
+    },
+    {
+      "key": "dc.date.available",
+      "value": "2012-10-01T18:08:11Z",
+      "language": null
+    },
+    {
+      "key": "dc.date.issued",
+      "value": "2007",
+      "language": null
+    },
+    {
+      "key": "dc.identifier.isbn",
+      "value": "5738002563",
+      "language": null
+    },
+    {
+      "key": "dc.identifier.uri",
+      "value": "http://arks.princeton.edu/ark:/88435/dsp012801pg38m",
+      "language": null
+    },
+    {
+      "key": "dc.publisher",
+      "value": "Institute of Tolerance; UNESCO Moscow; Rudomino Publishers;",
+      "language": "en_US"
+    },
+    {
+      "key": "dc.subject",
+      "value": "Food",
+      "language": "en_US"
+    },
+    {
+      "key": "dc.subject",
+      "value": "Cuisine",
+      "language": "en_US"
+    },
+    {
+      "key": "dc.title",
+      "value": "Travelling across dinner-tables.",
+      "language": "en_US"
+    },
+    {
+      "key": "pu.projectgrantnumber",
+      "value": "690-1011",
+      "language": "en_US"
+    },
+    {
+      "key": "pu.depositor",
+      "value": "Pendse, Liladhar",
+      "language": null
+    }
+  ],
+  "bitstreams": [
+    {
+      "id": 3927,
+      "name": "dinner tables.pdf",
+      "handle": null,
+      "type": "bitstream",
+      "link": "/rest/bitstreams/3927",
+      "expand": [
+        "parent",
+        "policies",
+        "all"
+      ],
+      "bundleName": "ORIGINAL",
+      "description": null,
+      "format": "Adobe PDF",
+      "mimeType": "application/pdf",
+      "sizeBytes": 2065179,
+      "parentObject": null,
+      "retrieveLink": "/bitstreams/3927/retrieve",
+      "checkSum": {
+        "value": "e759a5f4521c8c68361be345acef9983",
+        "checkSumAlgorithm": "MD5"
+      },
+      "sequenceId": 1,
+      "policies": null
+    },
+    {
+      "id": 3928,
+      "name": "license.txt",
+      "handle": null,
+      "type": "bitstream",
+      "link": "/rest/bitstreams/3928",
+      "expand": [
+        "parent",
+        "policies",
+        "all"
+      ],
+      "bundleName": "LICENSE",
+      "description": null,
+      "format": "License",
+      "mimeType": "text/plain",
+      "sizeBytes": 677,
+      "parentObject": null,
+      "retrieveLink": "/bitstreams/3928/retrieve",
+      "checkSum": {
+        "value": "eed0bb67d228ded438310bb3a7b0d0f5",
+        "checkSumAlgorithm": "MD5"
+      },
+      "sequenceId": 2,
+      "policies": null
+    },
+    {
+      "id": 5127,
+      "name": "dinner tables.pdf.txt",
+      "handle": null,
+      "type": "bitstream",
+      "link": "/rest/bitstreams/5127",
+      "expand": [
+        "parent",
+        "policies",
+        "all"
+      ],
+      "bundleName": "TEXT",
+      "description": "Extracted text",
+      "format": "Text",
+      "mimeType": "text/plain",
+      "sizeBytes": 72974,
+      "parentObject": null,
+      "retrieveLink": "/bitstreams/5127/retrieve",
+      "checkSum": {
+        "value": "9c69494781341971b785729e4870ed48",
+        "checkSumAlgorithm": "MD5"
+      },
+      "sequenceId": 3,
+      "policies": null
+    }
+  ],
+  "archived": "true",
+  "withdrawn": "false"
+}
