+++ spec/fixtures/dspace/monographic_collection.json
@@ -1,0 +1,103 @@
+{
+  "id": 222,
+  "name": "Monographic reports and papers (Publicly Accessible)",
+  "handle": "88435/dsp016q182k16g",
+  "type": "collection",
+  "link": "/rest/collections/222",
+  "expand": [],
+  "logo": null,
+  "parentCommunity": {
+    "id": 146,
+    "name": "Princeton University Library",
+    "handle": "88435/dsp01g732d9019",
+    "type": "community",
+    "link": "/rest/communities/146",
+    "expand": [
+      "parentCommunity",
+      "collections",
+      "subCommunities",
+      "logo",
+      "all"
+    ],
+    "logo": null,
+    "parentCommunity": null,
+    "copyrightText": "",
+    "introductoryText": "",
+    "shortDescription": "",
+    "sidebarText": "",
+    "countItems": 12954,
+    "subcommunities": [],
+    "collections": []
+  },
+  "parentCommunityList": [
+    {
+      "id": 146,
+      "name": "Princeton University Library",
+      "handle": "88435/dsp01g732d9019",
+      "type": "community",
+      "link": "/rest/communities/146",
+      "expand": [
+        "parentCommunity",
+        "collections",
+        "subCommunities",
+        "logo",
+        "all"
+      ],
+      "logo": null,
+      "parentCommunity": null,
+      "copyrightText": "",
+      "introductoryText": "",
+      "shortDescription": "",
+      "sidebarText": "",
+      "countItems": 12954,
+      "subcommunities": [],
+      "collections": []
+    }
+  ],
+  "items": [
+    {
+      "id": 1672,
+      "name": "Travelling across dinner-tables.",
+      "handle": "88435/dsp012801pg38m",
+      "type": "item",
+      "link": "/rest/items/1672",
+      "expand": [
+        "metadata",
+        "parentCollection",
+        "parentCollectionList",
+        "parentCommunityList",
+        "bitstreams",
+        "all"
+      ],
+      "lastModified": "2012-10-07 02:01:57.234",
+      "parentCollection": null,
+      "parentCollectionList": null,
+      "parentCommunityList": null,
+      "bitstreams": null,
+      "archived": "true",
+      "withdrawn": "false"
+    },
+    {
+      "id": 93362,
+      "name": "Recenseamento do Brazil em 1872",
+      "handle": "88435/dsp01h989r5980",
+      "type": "item",
+      "link": "/rest/items/93362",
+      "expand": [],
+      "lastModified": "2018-11-10 04:19:33.258",
+      "parentCollection": null,
+      "parentCollectionList": null,
+      "parentCommunityList": null,
+      "metadata": null,
+      "bitstreams": null,
+      "archived": "true",
+      "withdrawn": "false"
+    }
+  ],
+  "license": "<p>By agreeing, you are verifying that, to the best of your knowledge,<br>\r\n1. this file meets the guidelines for these collections\r\nas expressed on the Collection Development Sharepoint <a href=\"https://sp.princeton.edu/library/selectors/DSpacePDFLibrary/default.aspx\">DSPACE Digital Publications</a> site; and that<br>\r\n2. if the file is to be publicly available, its content is not restricted by copyright regulations; and also that<br>\r\n3. if access is to be limited to the Princeton Network, it meets Princeton’s license/publishers agreements<br>\r\nQuestions regarding the guidelines may be directed to <a href=\"mailto:pattygb@princeton.edu\">pattygb@princeton.edu</a></p>",
+  "copyrightText": "",
+  "introductoryText": "",
+  "shortDescription": "",
+  "sidebarText": "",
+  "numberItems": 1982
+}
